Task #81, PHP Session
I added a PHP session to each HTML.  As such, it had to be converted to PHP.  So, all of the old HTML is in static/unused/html.  If you start at index and go to result, it'll show something to screen.  I think I can get a better result, so I'll work on that.